Added turnbased support for RLikes
Created a "TurnBasedGameManager" which manages two teams (A,B) and waits
for all members of a team to act before proceeding to next turn.

Create a Base Rogue Like (BaseRL) AI type which moves directly to the
player.

Tested on square (4 way and 8 way) and Hex grids.

Known issue: The turn based manager can't proceed to next turn when
enemy is directly adjacent because it will never act. This will be fixed
once I implement attacking.